+++ src/parser/index.ts
@@ -5,5 +5,22 @@
  */
 
 export * from './ast';
+
+export { ExpressionParser } from './expression-parser';
 export { Lexer } from './lexer';
-export { ParseError, type ParseResult, Parser } from './parser';
+export { type ParseResult, Parser } from './parser';
+export {
+  MAX_RECURSION_DEPTH,
+  MAX_TOKEN_COUNT,
+  ParseError,
+  ParserBase,
+  TYPE_KEYWORDS,
+} from './parser-base';
+
+export {
+  KEYWORDS,
+  OPERATORS,
+  SORTED_SYMBOL_OPERATORS,
+  type Token,
+  TokenType,
+} from './token-types';